+++ Sources/StreamChat/APIClient/CDNClient.swift
@@ -61,7 +61,7 @@ class StreamCDNClient: CDNClient {
             fileName: uploadingState.localFileURL.lastPathComponent,
             mimeType: uploadingState.file.type.mimeType
         )
-        let endpoint = Endpoint<FileUploadPayload>.uploadAttachment(with: attachment.id, type: attachment.type)
+        let endpoint = Endpoint<FileUploadPayload>.uploadAttachment(with: attachment.id.cid, type: attachment.type)
         
         encoder.encodeRequest(for: endpoint) { [weak self] (requestResult) in
             var urlRequest: URLRequest