+++ Sources/StreamChat/WebSocketClient/EventMiddlewares/ChannelReadUpdaterMiddleware.swift
@@ -11,16 +11,14 @@ struct ChannelReadUpdaterMiddleware: EventMiddleware {
         case let event as MessageNewEventDTO:
             increaseUnreadCountIfNeeded(
                 for: event.cid,
-                userId: event.user.id,
-                newMessageAt: event.createdAt,
+                message: event.message,
                 session: session
             )
-            
+
         case let event as NotificationMessageNewEventDTO:
             increaseUnreadCountIfNeeded(
                 for: event.channel.cid,
-                userId: event.message.user.id,
-                newMessageAt: event.createdAt,
+                message: event.message,
                 session: session
             )
             
@@ -74,18 +72,27 @@ struct ChannelReadUpdaterMiddleware: EventMiddleware {
     
     private func increaseUnreadCountIfNeeded(
         for cid: ChannelId,
-        userId: UserId,
-        newMessageAt: Date,
+        message: MessagePayload,
         session: DatabaseSession
     ) {
-        guard let currentUserId = session.currentUser?.user.id, currentUserId != userId else {
-            // Own messages don't increase unread count
+        // Silent messages don't increase unread count
+        if message.isSilent {
             return
         }
-        
-        // Try to get the existing channel read for the CURRENT user
+
+        // Thread replies don't increase unread count
+        if message.parentId != nil && !message.showReplyInChannel {
+            return
+        }
+
+        // Own messages don't increase unread count
+        guard let currentUserId = session.currentUser?.user.id, currentUserId != message.user.id else {
+            return
+        }
+
+        // Try to get the existing channel read for the current user
         if let read = session.loadChannelRead(cid: cid, userId: currentUserId) {
-            if newMessageAt > read.lastReadAt {
+            if message.createdAt > read.lastReadAt {
                 read.unreadMessageCount += 1
             }
         } else {